feat: add prevent delete data if has related data

+++ app/Http/Controllers/Admin/SubDistrictController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreSubDistrictRequest;
 use App\Http\Requests\UpdateSubDistrictRequest;
 use App\Models\SubDistrict;
+use App\Models\TouristDestination;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
@@ -99,6 +100,12 @@ class SubDistrictController extends Controller
     {
         abort_if(! auth()->user()->is_admin, 403);
 
+        if ($subDistrict->loadCount(['touristDestinations'])->tourist_destinations_count > 0) {
+            toastr()->warning('Terdapat data destinasi wisata, hapus atau ubah terlebih dahulu data destinasi wisata terkait', 'Data Tidak Dapat Dihapus');
+
+            return redirect(route('dashboard.sub-districts.related-tourist-destination', ['sub_district' => $subDistrict]));
+        }
+
         if ($subDistrict->geojson_path != null) {
             Storage::delete($subDistrict->geojson_path);
         }
@@ -112,5 +119,18 @@ class SubDistrictController extends Controller
     public function download(SubDistrict $subDistrict)
     {
         return Storage::download($subDistrict->geojson_path);
+    }
+
+    public function relatedTouristDestination(SubDistrict $subDistrict)
+    {
+        $touristDestinations = TouristDestination::select('slug', 'name', 'address', 'manager', 'distance_from_city_center', 'latitude', 'longitude')
+        ->where('sub_district_id', $subDistrict->id)
+            ->orderBy('name', 'asc');
+
+        return view('sub-district.related-tourist-destination', [
+            'touristDestinations' => $touristDestinations->paginate(10),
+            'touristDestinationMapping' => $touristDestinations->get(),
+            'subDistrict' => $subDistrict,
+        ]);
     }
 }
